fix bug where seam ix could become negative
np.argmin returns first occurrence of a value when duplicates
exist. when middle = 0, row[left] and row[middle] are
duplicates due to the ix clipping on the forward pass. thus,
when middle = 0, a negative seam index was being calculated.
this was allowing the seam to wrap around the image from left
to right.